Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -10,6 +10,6 @@ int main() {
    scanf("%i",&b);
    scanf("%i",&c);
    int sum = a+b+c;
-   int avg = sum / 3;
-   printf("Average %d",avg);
+   float avg = sum / 3;
+   printf("Average %.2f",avg);
 }
